requirements.txt
instaloader==4.14.1
Flask==3.1.1
requests==2.32.4
python-dotenv==1.1.1
Flask-SQLAlchemy==3.1.1
psycopg2==2.9.10